+++ src/main/java/WayofTime/bloodmagic/item/sigil/ItemSigilSuppression.java
@@ -47,5 +47,4 @@ public class ItemSigilSuppression extends ItemSigilToggleableBase {
             }
         }
     }
-
 }
