+++ src/main/java/WayofTime/bloodmagic/item/sigil/ItemSigilToggleable.java
@@ -1,4 +1,4 @@
-package WayofTime.bloodmagic.item;
+package WayofTime.bloodmagic.item.sigil;
 
 import WayofTime.bloodmagic.core.data.Binding;
 import WayofTime.bloodmagic.core.data.SoulTicket;
@@ -54,7 +54,7 @@ public class ItemSigilToggleable extends ItemSigil implements IActivatable {
         if (!world.isRemote && !isUnusable(stack)) {
             if (player.isSneaking())
                 setActivatedState(stack, !getActivated(stack));
-            if (getActivated(stack) && NetworkHelper.getSoulNetwork(player).syphonAndDamage(player, SoulTicket.item(stack, world, player, getLpUsed())).isSuccess())
+            if (getActivated(stack))
                 return super.onItemRightClick(world, player, hand);
         }
 
@@ -71,7 +71,7 @@ public class ItemSigilToggleable extends ItemSigil implements IActivatable {
         if (binding == null || player.isSneaking()) // Make sure Sigils are bound before handling. Also ignores while toggling state
             return EnumActionResult.PASS;
 
-        return (NetworkHelper.getSoulNetwork(binding).syphonAndDamage(player, SoulTicket.item(stack, world, player, getLpUsed())).isSuccess() && onSigilUse(player.getHeldItem(hand), player, world, pos, side, hitX, hitY, hitZ)) ? EnumActionResult.SUCCESS : EnumActionResult.FAIL;
+        return onSigilUse(player.getHeldItem(hand), player, world, pos, side, hitX, hitY, hitZ) ? EnumActionResult.SUCCESS : EnumActionResult.FAIL;
     }
 
     public boolean onSigilUse(ItemStack itemStack, EntityPlayer player, World world, BlockPos blockPos, EnumFacing side, float hitX, float hitY, float hitZ) {
